Fixed labelling so off-screen atoms / measurement points don't get drawn. Fixed link to Probe Atom on context menu.

+++ src/model/model.h
@@ -363,7 +363,7 @@ class Model
 	// Project the model coordinates of the atom into local and 2D coordinates
 	void projectAtom(Atom*);
 	// Project given model coordinates into screen coordinates
-	Vec3<int> &modelToScreen(Vec3<double> &pos);
+	Vec3<double> &modelToScreen(Vec3<double> &pos);
 	// Project the model coordinates of all atoms
 	void projectAll();
 	// Project the model coordinates of selected atoms
